fix(genai): fix tuning examples test issues
- Updates tuning file names to fix cicd
  - word `test` is reserved for test files
- Update region tag and file names.
  - add word `job` to filenames & region tags
