+++ genai/tuning/tuning_job_create.py
@@ -13,16 +13,32 @@
 # limitations under the License.
 
 
-def get_tuning_job(name: str) -> str:
-    # [START googlegenaisdk_tuning_get]
+def create_tuning_job() -> str:
+    # [START googlegenaisdk_tuning_job_create]
+    import time
+
     from google import genai
-    from google.genai.types import HttpOptions
+    from google.genai.types import HttpOptions, CreateTuningJobConfig
 
     client = genai.Client(http_options=HttpOptions(api_version="v1"))
 
-    # Get the tuning job and the tuned model.
-    # Eg. name = "projects/123456789012/locations/us-central1/tuningJobs/123456789012345"
-    tuning_job = client.tunings.get(name=name)
+    tuning_job = client.tunings.tune(
+        base_model="gemini-2.0-flash-lite-001",
+        training_dataset="gs://cloud-samples-data/ai-platform/generative_ai/gemini-2_0/text/sft_train_data.jsonl",
+        config=CreateTuningJobConfig(
+            tuned_model_display_name="Example tuning job",
+        ),
+    )
+
+    running_states = set([
+        "JOB_STATE_PENDING",
+        "JOB_STATE_RUNNING",
+    ])
+
+    while tuning_job.state in running_states:
+        print(tuning_job.state)
+        tuning_job = client.tunings.get(name=tuning_job.name)
+        time.sleep(60)
 
     print(tuning_job.tuned_model.model)
     print(tuning_job.tuned_model.endpoint)
@@ -32,10 +48,9 @@ def get_tuning_job(name: str) -> str:
     # projects/123456789012/locations/us-central1/endpoints/123456789012345
     # projects/123456789012/locations/us-central1/metadataStores/default/contexts/tuning-experiment-2025010112345678
 
-    # [END googlegenaisdk_tuning_get]
+    # [END googlegenaisdk_tuning_job_create]
     return tuning_job.name
 
 
 if __name__ == "__main__":
-    tuning_job_name = input("Tuning job name: ")
-    get_tuning_job(tuning_job_name)
+    create_tuning_job()
