+++ genai/tuning/tuning_job_get.py
@@ -1,4 +1,4 @@
-# Copyright 2024 Google LLC
+# Copyright 2025 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,8 +13,8 @@
 # limitations under the License.
 
 
-def test_tuned_endpoint(name: str) -> str:
-    # [START googlegenaisdk_tuning_test]
+def get_tuning_job(name: str) -> str:
+    # [START googlegenaisdk_tuning_job_get]
     from google import genai
     from google.genai.types import HttpOptions
 
@@ -24,19 +24,18 @@ def test_tuned_endpoint(name: str) -> str:
     # Eg. name = "projects/123456789012/locations/us-central1/tuningJobs/123456789012345"
     tuning_job = client.tunings.get(name=name)
 
-    contents = "Why is the sky blue?"
+    print(tuning_job.tuned_model.model)
+    print(tuning_job.tuned_model.endpoint)
+    print(tuning_job.experiment)
+    # Example response:
+    # projects/123456789012/locations/us-central1/models/1234567890@1
+    # projects/123456789012/locations/us-central1/endpoints/123456789012345
+    # projects/123456789012/locations/us-central1/metadataStores/default/contexts/tuning-experiment-2025010112345678
 
-    # Tests the default checkpoint
-    response = client.models.generate_content(
-        model=tuning_job.tuned_model.endpoint,
-        contents=contents,
-    )
-    print(response.text)
-
-    # [END googlegenaisdk_tuning_test]
-    return response.text
+    # [END googlegenaisdk_tuning_job_get]
+    return tuning_job.name
 
 
 if __name__ == "__main__":
     tuning_job_name = input("Tuning job name: ")
-    test_tuned_endpoint(tuning_job_name)
+    get_tuning_job(tuning_job_name)
